My change

+++ backend/models.py
@@ -1,6 +1,5 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean, Text, ForeignKey, Numeric
+from sqlalchemy import Column, Integer, String, Float, DateTime, Boolean, Text, ForeignKey
 from sqlalchemy.orm import relationship
-from sqlalchemy.sql import func
 from database import Base
 from datetime import datetime
 
@@ -22,7 +21,7 @@ class Member(Base):
     phone_number = Column(String(20))
     address = Column(Text)
     join_date = Column(String(20))
-    monthly_sandha_amount = Column(Numeric(12, 2), default=1000)
+    monthly_sandha_amount = Column(Float, default=1000.0)
     status = Column(String(20), default="active")
     created_at = Column(DateTime, default=datetime.utcnow)
 
@@ -31,12 +30,12 @@ class Loan(Base):
     id = Column(Integer, primary_key=True)
     loan_number = Column(String(20), unique=True, index=True)
     member_id = Column(Integer, ForeignKey("members.id"))
-    principal_amount = Column(Numeric(12, 2))
+    principal_amount = Column(Float)
     interest_rate = Column(Float, default=1.0)
     loan_type = Column(String(20))
     loan_status = Column(String(20), default="active")
-    outstanding_principal = Column(Numeric(12, 2))
-    outstanding_interest = Column(Numeric(12, 2), default=0)
+    outstanding_principal = Column(Float)
+    outstanding_interest = Column(Float, default=0.0)
     created_at = Column(DateTime, default=datetime.utcnow)
 
 class MemberContribution(Base):
@@ -44,7 +43,7 @@ class MemberContribution(Base):
     id = Column(Integer, primary_key=True)
     member_id = Column(Integer, ForeignKey("members.id"))
     contribution_month = Column(String(20))
-    amount = Column(Numeric(12, 2))
+    amount = Column(Float)
     payment_status = Column(String(20), default="pending")
     created_at = Column(DateTime, default=datetime.utcnow)
 
@@ -53,15 +52,15 @@ class LoanRepayment(Base):
     id = Column(Integer, primary_key=True)
     loan_id = Column(Integer, ForeignKey("loans.id"))
     payment_date = Column(String(20))
-    principal_amount = Column(Numeric(12, 2))
-    interest_amount = Column(Numeric(12, 2))
-    total_amount = Column(Numeric(12, 2))
+    principal_amount = Column(Float)
+    interest_amount = Column(Float)
+    total_amount = Column(Float)
     created_at = Column(DateTime, default=datetime.utcnow)
 
 class Settings(Base):
     __tablename__ = "settings"
     id = Column(Integer, primary_key=True)
-    default_sandha = Column(Numeric(12, 2), default=1000)
+    default_sandha = Column(Float, default=1000.0)
     default_interest_rate = Column(Float, default=1.0)
     monthly_repayment_percentage = Column(Float, default=10.0)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
